[refactor] 소캣 리팩토링 수행 #74
- back-end 소켓의 코드를 분리함.
